backend/requirements.txt
fastapi[standard]
pydantic
sqlalchemy
python-dotenv
uvicorn
asyncpg
fastapi-cache2
alembic
psycopg2-binary
redis